Change sign for diffusion term and create plots
The paper has a negative sign in front of the diffusion term for the equation, but I don't think this makes sense. Maybe it depends on the context - negative diffusion does seem to be possible. For our purposes, I've changed it to positive to correspond to the case of dye transport. Also added code for creating plots for a given timestep. It can be clearly seen that the dye diffuses and undergoes convection with the fluid.